+++ tests/testthat/test-plot_box.R
@@ -8,18 +8,18 @@ initialize_db()
 start_db()
 
 # enter data -------------------------------------------------------------------
-dbWriteTable(globaltrends_db, "data_score", data_score, append = TRUE)
-dbWriteTable(globaltrends_db, "data_doi", data_doi, append = TRUE)
+dbWriteTable(globaltrends_db, "data_score", example_score, append = TRUE)
+dbWriteTable(globaltrends_db, "data_doi", example_doi, append = TRUE)
 
 # plot_box.exp_score ------------------------------------------------------------
 test_that("plot_box.exp_score1", {
-  keywords <- unique(data_score$keyword)[1:9]
+  keywords <- unique(example_score$keyword)[1:9]
   data <- map_dfr(keywords, export_score) %>%
     filter(location == "US")
   out1 <- plot_box(data, type = "obs")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_score$keyword)[1:10]
+  keywords <- unique(example_score$keyword)[1:10]
   data <- map_dfr(keywords, export_score) %>%
     filter(location == "US")
   expect_warning(
@@ -32,13 +32,13 @@ test_that("plot_box.exp_score1", {
 })
 
 test_that("plot_box.exp_score2", {
-  keywords <- unique(data_score$keyword)[1:9]
+  keywords <- unique(example_score$keyword)[1:9]
   data <- map_dfr(keywords, export_score) %>%
     filter(location == "US")
   out1 <- plot_box(data, type = "sad")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_score$keyword)[1:10]
+  keywords <- unique(example_score$keyword)[1:10]
   data <- map_dfr(keywords, export_score) %>%
     filter(location == "US")
   expect_warning(
@@ -51,13 +51,13 @@ test_that("plot_box.exp_score2", {
 })
 
 test_that("plot_box.exp_score3", {
-  keywords <- unique(data_score$keyword)[1:9]
+  keywords <- unique(example_score$keyword)[1:9]
   data <- map_dfr(keywords, export_score) %>%
     filter(location == "US")
   out1 <- plot_box(data, type = "trd")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_score$keyword)[1:10]
+  keywords <- unique(example_score$keyword)[1:10]
   data <- map_dfr(keywords, export_score) %>%
     filter(location == "US")
   expect_warning(
@@ -70,12 +70,12 @@ test_that("plot_box.exp_score3", {
 })
 
 test_that("plot_box.exp_score4", {
-  data <- export_score(keyword = data_score$keyword[[1]]) %>%
+  data <- export_score(keyword = example_score$keyword[[1]]) %>%
     filter(location == "CN")
   out1 <- plot_box(data, type = "obs")
   expect_s3_class(out1, "ggplot")
 
-  data <- export_score(keyword = data_score$keyword[[1]])
+  data <- export_score(keyword = example_score$keyword[[1]])
   expect_warning(
     out2 <- plot_box(data, type = "obs"),
     "The plot function is limited to 1 location\\.\nYou use 3 locations\\.\nOnly 'CN' is used\\."
@@ -87,7 +87,7 @@ test_that("plot_box.exp_score4", {
 
 # plot_box.exp_score defaults ---------------------------------------------------
 test_that("plot_box.exp_score5", {
-  keywords <- unique(data_score$keyword)[1:9]
+  keywords <- unique(example_score$keyword)[1:9]
   data <- map_dfr(keywords, export_score) %>%
     filter(location == "CN")
   out1 <- plot_box(data)
@@ -157,12 +157,12 @@ test_that("plot_box.abnorm_score1", {
 })
 
 test_that("plot_box.abnorm_score2", {
-  data <- export_score(keyword = unique(data_score$keyword)[[1]], location = "US")
+  data <- export_score(keyword = unique(example_score$keyword)[[1]], location = "US")
   data <- get_abnorm_hist(data)
   out1 <- plot_box(data)
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_score$keyword)[1:2]
+  keywords <- unique(example_score$keyword)[1:2]
   data <- map_dfr(keywords, export_score, location = "US")
   data <- get_abnorm_hist(data)
   expect_warning(
@@ -175,12 +175,12 @@ test_that("plot_box.abnorm_score2", {
 })
 
 test_that("plot_box.abnorm_score3", {
-  data <- export_score(keyword = unique(data_score$keyword)[[1]], location = "CN")
+  data <- export_score(keyword = unique(example_score$keyword)[[1]], location = "CN")
   data <- get_abnorm_hist(data)
   out1 <- plot_box(data)
   expect_s3_class(out1, "ggplot")
 
-  data <- export_score(keyword = unique(data_score$keyword)[[1]])
+  data <- export_score(keyword = unique(example_score$keyword)[[1]])
   data <- get_abnorm_hist(data)
   expect_warning(
     out2 <- plot_box(data),
@@ -193,12 +193,12 @@ test_that("plot_box.abnorm_score3", {
 
 # plot_box.exp_voi --------------------------------------------------------------
 test_that("plot_box.exp_voi1", {
-  keywords <- unique(data_score$keyword)[1:9]
+  keywords <- unique(example_score$keyword)[1:9]
   data <- map_dfr(keywords, export_voi)
   out1 <- plot_box(data, type = "obs")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_score$keyword)[1:10]
+  keywords <- unique(example_score$keyword)[1:10]
   data <- map_dfr(keywords, export_voi)
   expect_warning(
     out2 <- plot_box(data, type = "obs"),
@@ -210,12 +210,12 @@ test_that("plot_box.exp_voi1", {
 })
 
 test_that("plot_box.exp_voi2", {
-  keywords <- unique(data_score$keyword)[1:9]
+  keywords <- unique(example_score$keyword)[1:9]
   data <- map_dfr(keywords, export_voi)
   out1 <- plot_box(data, type = "sad")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_score$keyword)[1:10]
+  keywords <- unique(example_score$keyword)[1:10]
   data <- map_dfr(keywords, export_voi)
   expect_warning(
     out2 <- plot_box(data, type = "sad"),
@@ -227,12 +227,12 @@ test_that("plot_box.exp_voi2", {
 })
 
 test_that("plot_box.exp_voi3", {
-  keywords <- unique(data_score$keyword)[1:9]
+  keywords <- unique(example_score$keyword)[1:9]
   data <- map_dfr(keywords, export_voi)
   out1 <- plot_box(data, type = "trd")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_score$keyword)[1:10]
+  keywords <- unique(example_score$keyword)[1:10]
   data <- map_dfr(keywords, export_voi)
   expect_warning(
     out2 <- plot_box(data, type = "trd"),
@@ -245,7 +245,7 @@ test_that("plot_box.exp_voi3", {
 
 # plot_box.exp_voi defaults -----------------------------------------------------
 test_that("plot_box.exp_voi4", {
-  keywords <- unique(data_score$keyword)[1:9]
+  keywords <- unique(example_score$keyword)[1:9]
   data <- map_dfr(keywords, export_voi)
   out1 <- plot_box(data)
   out2 <- plot_box(data, type = "obs")
@@ -314,12 +314,12 @@ test_that("plot_box.abnorm_voi1", {
 })
 
 test_that("plot_box.abnorm_voi2", {
-  data <- export_voi(keyword = unique(data_score$keyword)[[1]])
+  data <- export_voi(keyword = unique(example_score$keyword)[[1]])
   data <- get_abnorm_hist(data)
   out1 <- plot_box(data)
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_score$keyword)[1:2]
+  keywords <- unique(example_score$keyword)[1:2]
   data <- map_dfr(keywords, export_voi)
   data <- get_abnorm_hist(data)
   expect_warning(
@@ -333,12 +333,12 @@ test_that("plot_box.abnorm_voi2", {
 
 # plot_box.exp_doi gini ---------------------------------------------------------
 test_that("plot_box.exp_doi1a", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data, type = "obs", measure = "gini")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_doi$keyword)[1:10]
+  keywords <- unique(example_doi$keyword)[1:10]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   expect_warning(
     out2 <- plot_box(data, type = "obs", measure = "gini"),
@@ -350,12 +350,12 @@ test_that("plot_box.exp_doi1a", {
 })
 
 test_that("plot_box.exp_doi2a", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data, type = "sad", measure = "gini")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_doi$keyword)[1:10]
+  keywords <- unique(example_doi$keyword)[1:10]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   expect_warning(
     out2 <- plot_box(data, type = "sad", measure = "gini"),
@@ -367,12 +367,12 @@ test_that("plot_box.exp_doi2a", {
 })
 
 test_that("plot_box.exp_doi3a", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data, type = "trd", measure = "gini")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_doi$keyword)[1:10]
+  keywords <- unique(example_doi$keyword)[1:10]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   expect_warning(
     out2 <- plot_box(data, type = "trd", measure = "gini"),
@@ -385,12 +385,12 @@ test_that("plot_box.exp_doi3a", {
 
 # plot_box.exp_doi hhi ----------------------------------------------------------
 test_that("plot_box.exp_doi1b", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data, type = "obs", measure = "hhi")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_doi$keyword)[1:10]
+  keywords <- unique(example_doi$keyword)[1:10]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   expect_warning(
     out2 <- plot_box(data, type = "obs", measure = "hhi"),
@@ -402,12 +402,12 @@ test_that("plot_box.exp_doi1b", {
 })
 
 test_that("plot_box.exp_doi2b", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data, type = "sad", measure = "hhi")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_doi$keyword)[1:10]
+  keywords <- unique(example_doi$keyword)[1:10]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   expect_warning(
     out2 <- plot_box(data, type = "sad", measure = "hhi"),
@@ -419,12 +419,12 @@ test_that("plot_box.exp_doi2b", {
 })
 
 test_that("plot_box.exp_doi3b", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data, type = "trd", measure = "hhi")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_doi$keyword)[1:10]
+  keywords <- unique(example_doi$keyword)[1:10]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   expect_warning(
     out2 <- plot_box(data, type = "trd", measure = "hhi"),
@@ -437,12 +437,12 @@ test_that("plot_box.exp_doi3b", {
 
 # plot_box.exp_doi entropy ------------------------------------------------------
 test_that("plot_box.exp_doi1c", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data, type = "obs", measure = "entropy")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_doi$keyword)[1:10]
+  keywords <- unique(example_doi$keyword)[1:10]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   expect_warning(
     out2 <- plot_box(data, type = "obs", measure = "entropy"),
@@ -454,12 +454,12 @@ test_that("plot_box.exp_doi1c", {
 })
 
 test_that("plot_box.exp_doi2c", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data, type = "sad", measure = "entropy")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_doi$keyword)[1:10]
+  keywords <- unique(example_doi$keyword)[1:10]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   expect_warning(
     out2 <- plot_box(data, type = "sad", measure = "entropy"),
@@ -471,12 +471,12 @@ test_that("plot_box.exp_doi2c", {
 })
 
 test_that("plot_box3c", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data, type = "trd", measure = "entropy")
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_doi$keyword)[1:10]
+  keywords <- unique(example_doi$keyword)[1:10]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   expect_warning(
     out2 <- plot_box(data, type = "trd", measure = "entropy"),
@@ -489,7 +489,7 @@ test_that("plot_box3c", {
 
 # plot_box.exp_doi defaults -----------------------------------------------------
 test_that("plot_box.exp_doi4a", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data = data, measure = "gini")
   out2 <- plot_box(data = data, measure = "gini", type = "obs")
@@ -503,7 +503,7 @@ test_that("plot_box.exp_doi4a", {
 })
 
 test_that("plot_box.exp_doi4b", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data = data, measure = "hhi")
   out2 <- plot_box(data = data, measure = "hhi", type = "obs")
@@ -517,7 +517,7 @@ test_that("plot_box.exp_doi4b", {
 })
 
 test_that("plot_box.exp_doi4c", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   out1 <- plot_box(data = data, measure = "entropy")
   out2 <- plot_box(data = data, measure = "entropy", type = "obs")
@@ -531,7 +531,7 @@ test_that("plot_box.exp_doi4c", {
 })
 
 test_that("plot_box.exp_doi4d", {
-  keywords <- unique(data_doi$keyword)[1:9]
+  keywords <- unique(example_doi$keyword)[1:9]
   data <- map_dfr(keywords, export_doi, type = "obs", locations = "countries")
   out1 <- plot_box(data = data)
   out2 <- plot_box(data = data, measure = "gini")
@@ -644,12 +644,12 @@ test_that("plot_box.abnorm_doi1", {
 })
 
 test_that("plot_box.abnorm_doi2", {
-  data <- export_doi(keyword = unique(data_doi$keyword)[[1]], locations = "countries")
+  data <- export_doi(keyword = unique(example_doi$keyword)[[1]], locations = "countries")
   data <- get_abnorm_hist(data)
   out1 <- plot_box(data)
   expect_s3_class(out1, "ggplot")
 
-  keywords <- unique(data_doi$keyword)[1:2]
+  keywords <- unique(example_doi$keyword)[1:2]
   data <- map_dfr(keywords, export_doi, locations = "countries")
   data <- get_abnorm_hist(data)
   expect_warning(